# BooksLoan/BooksLoan/OpenAPIs/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Book": {
      "get": {
        "tags": [
          "Book"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      }
    },
    "/api/Book/{id}": {
      "get": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Rating": {
      "get": {
        "tags": [
          "Rating"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rating"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rating"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rating"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rating"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              }
            }
          }
        }
      }
    },
    "/api/Rating/{id}": {
      "get": {
        "tags": [
          "Rating"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Rating"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Rating"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Rating"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Author": {
        "required": [
          "createdBy",
          "creationDate",
          "firstName",
          "isActive",
          "lastModificationDate",
          "lastModifiedBy",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastModificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "maxLength": 150,
            "type": "string"
          },
          "lastName": {
            "maxLength": 150,
            "type": "string"
          },
          "middleName": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "pseudonym": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "dob": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "nationality": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "book": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Book": {
        "required": [
          "createdBy",
          "creationDate",
          "isActive",
          "isbn",
          "lastModificationDate",
          "lastModifiedBy",
          "publishDate",
          "publisher",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastModificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 50,
            "type": "string"
          },
          "subTitle": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "orginalTitle": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "isbn": {
            "maxLength": 50,
            "type": "string"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "publisher": {
            "maxLength": 50,
            "type": "string"
          },
          "language": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "rating": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Category": {
        "required": [
          "createdBy",
          "creationDate",
          "isActive",
          "lastModificationDate",
          "lastModifiedBy",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastModificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "type": "string"
          },
          "book": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Loan": {
        "required": [
          "createdBy",
          "creationDate",
          "isActive",
          "lastModificationDate",
          "lastModifiedBy",
          "loanDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastModificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "loanDate": {
            "maxLength": 150,
            "type": "string",
            "format": "date-time"
          },
          "returnDate": {
            "maxLength": 150,
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "freeDate": {
            "maxLength": 150,
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "readerId": {
            "type": "integer",
            "format": "int32"
          },
          "reader": {
            "$ref": "#/components/schemas/Reader"
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          }
        },
        "additionalProperties": false
      },
      "Rating": {
        "required": [
          "createdBy",
          "creationDate",
          "isActive",
          "lastModificationDate",
          "lastModifiedBy",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastModificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "maxLength": 50,
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "readerId": {
            "type": "integer",
            "format": "int32"
          },
          "reader": {
            "$ref": "#/components/schemas/Reader"
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          }
        },
        "additionalProperties": false
      },
      "Reader": {
        "required": [
          "createdBy",
          "creationDate",
          "email",
          "firstName",
          "isActive",
          "lastModificationDate",
          "lastModifiedBy",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastModificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "maxLength": 150,
            "type": "string"
          },
          "lastName": {
            "maxLength": 150,
            "type": "string"
          },
          "middleName": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "email": {
            "maxLength": 150,
            "type": "string"
          },
          "nick": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "dob": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "nationality": {
            "maxLength": 150,
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            },
            "nullable": true
          },
          "loan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loan"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}